Change __blksize_t to be 64 bits
[ make this in sync with glibc ]

+++ kern/include/sys/types.h
@@ -16,7 +16,7 @@ typedef uint32_t __uid_t;
 typedef uint32_t __gid_t;
 typedef int64_t __off64_t;
 typedef __off64_t off64_t;
-typedef int32_t __blksize_t;
+typedef int64_t __blksize_t;
 typedef int64_t __blkcnt64_t;
 
 typedef uint64_t __le64;
